# packages/contracts/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "easyTree API",
    "version": "1.0.0",
    "description": "Vertrag zwischen Web-Shells und Backend. Erzeugt aus den Zod-Schemata in @easytree/contracts; nicht von Hand bearbeiten."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/kosten/mitarbeiter": {
      "get": {
        "operationId": "listEmployeesForRates",
        "summary": "Mitarbeiterliste fuer die Satzverwaltung",
        "responses": {
          "200": {
            "description": "Mitarbeiter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeesForRates"
                }
              }
            }
          },
          "400": {
            "description": "Ungueltige Anfrage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Nicht berechtigt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Konflikt oder veralteter Planstand",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        }
      }
    },
    "/kosten/stundensaetze/{employeeId}": {
      "get": {
        "operationId": "getRateHistory",
        "summary": "Aktive Satzversion und Historie eines Mitarbeiters",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Satzhistorie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateHistory"
                }
              }
            }
          },
          "400": {
            "description": "Ungueltige Anfrage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Nicht berechtigt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Konflikt oder veralteter Planstand",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        }
      }
    },
    "/kosten/stundensaetze": {
      "post": {
        "operationId": "createRateVersion",
        "summary": "Neue unveraenderliche Satzversion anlegen",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Mandantengebundener Idempotenzschluessel. Ein Wiederholungsaufruf liefert dieselbe Antwort, ohne den Effekt zu verdoppeln.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9._~-]+$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRateVersionCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Angelegte Version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateVersionDto"
                }
              }
            }
          },
          "400": {
            "description": "Ungueltige Anfrage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "401": {
            "description": "Nicht angemeldet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "Nicht berechtigt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "Konflikt oder veralteter Planstand",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDocument": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "correlationId"
        ],
        "additionalProperties": false
      },
...
      "EmployeesForRates": {
        "type": "object",
        "properties": {
          "employees": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "description": "UUID"
                },
                "displayName": {
                  "type": "string",
                  "minLength": 1
                },
                "active": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "displayName",
                "active"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "employees"
        ],
        "additionalProperties": false
      },
      "RateHistory": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          },
          "activeVersionId": {
            "nullable": true,
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          },
          "versions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "description": "UUID"
                },
                "employeeId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "description": "UUID"
                },
                "amountMinorUnits": {
                  "type": "string",
                  "maxLength": 18,
                  "pattern": "^\\d+$"
                },
                "currency": {
                  "type": "string",
                  "enum": [
                    "EUR"
                  ]
                },
                "validFrom": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                "validTo": {
                  "nullable": true,
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "aktiv",
                    "kommend",
                    "abgelaufen"
                  ]
                },
                "predecessorId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "description": "UUID"
                },
                "reason": {
                  "type": "string",
                  "minLength": 1
                },
                "createdAt": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
                  "description": "UTC-Zeitpunkt nach ISO 8601, drei Nachkommastellen, Suffix Z"
                },
                "createdBy": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "description": "UUID"
                }
              },
              "required": [
                "id",
                "employeeId",
                "amountMinorUnits",
                "currency",
                "validFrom",
                "validTo",
                "status",
                "predecessorId",
                "reason",
                "createdAt",
                "createdBy"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "employeeId",
          "activeVersionId",
          "versions"
        ],
        "additionalProperties": false
      },
      "RateVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          },
          "amountMinorUnits": {
            "type": "string",
            "maxLength": 18,
            "pattern": "^\\d+$"
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR"
            ]
          },
          "validFrom": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "validTo": {
            "nullable": true,
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "aktiv",
              "kommend",
              "abgelaufen"
            ]
          },
          "predecessorId": {
            "nullable": true,
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          },
          "reason": {
            "type": "string",
            "minLength": 1
          },
          "createdAt": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "description": "UTC-Zeitpunkt nach ISO 8601, drei Nachkommastellen, Suffix Z"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          }
        },
        "required": [
          "id",
          "employeeId",
          "amountMinorUnits",
          "currency",
          "validFrom",
          "validTo",
          "status",
          "predecessorId",
          "reason",
          "createdAt",
          "createdBy"
        ],
        "additionalProperties": false
      },
      "CreateRateVersionCommand": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          },
          "amountMinorUnits": {
            "type": "string",
            "maxLength": 18,
            "pattern": "^\\d+$"
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR"
            ]
          },
          "validFrom": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "validTo": {
            "nullable": true,
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "expectedActiveVersionId": {
            "nullable": true,
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "UUID"
          }
        },
        "required": [
          "employeeId",
          "amountMinorUnits",
          "currency",
          "validFrom",
          "validTo",
          "reason",
          "expectedActiveVersionId"
        ],
        "additionalProperties": false
      }
    }
  }
}